Remove Celery refs
